requirements.txt
Flask==2.0.1
google-generativeai
python-dotenv==0.19.0